+++ Assets/Scripts/DTOs/TestVerificationDTO.cs
@@ -20,14 +20,14 @@ public class TestVerificationDTO {
     public float CursorPositionY { get; set; }
     public float GazePositionX { get; set; }
     public float GazePositionY { get; set; }
-    public float? PupilDiameterLeft { get; set; }
-    public float? PupilDiameter3DLeft { get; set; }
-    public float? PupilDiameter3DRight { get; set; }
-    public float? PupilDiameterRight { get; set; }
-    public float? PupilConfidenceLeft { get; set; }
-    public float? PupilConfidenceRight { get; set; }
-    public float? PupilTimestampLeft { get; set; }
-    public float? PupilTimestampRight { get; set; }
+    public double? PupilDiameterLeft { get; set; }
+    public double? PupilDiameter3DLeft { get; set; }
+    public double? PupilDiameter3DRight { get; set; }
+    public double? PupilDiameterRight { get; set; }
+    public double? PupilConfidenceLeft { get; set; }
+    public double? PupilConfidenceRight { get; set; }
+    public double? PupilTimestampLeft { get; set; }
+    public double? PupilTimestampRight { get; set; }
     public float NosePositionX { get; set; }
     public float NosePositionY { get; set; }
     public float? HmdPositionX { get; set; }
@@ -67,9 +67,9 @@ public class TestVerificationDTO {
     /// <param name="hmdRotationZ"> Z coordinate rotation of the hmd.</param>
     public TestVerificationDTO( int testBlockId, string timeStamp, int targetIndex, float targetPositionX, float targetPositionY,
         float cursorPositionX, float cursorPositionY, float gazePositionX, float gazePositionY,
-        float? pupilDiameterLeft, float? pupilDiameterRight,
-        float? pupilDiameter3DLeft, float? pupilDiameter3DRight, float? pupilConfidenceLeft,
-        float? pupilConfidenceRight, float? pupilTimestampLeft, float? pupilTimestampRight,
+        double? pupilDiameterLeft, double? pupilDiameterRight,
+        double? pupilDiameter3DLeft, double? pupilDiameter3DRight, double? pupilConfidenceLeft,
+        double? pupilConfidenceRight, double? pupilTimestampLeft, double? pupilTimestampRight,
         float nosePositionX, float nosePositionY, float? hmdPositionX, float? hmdPositionY,
         float? hmdPositionZ, float? hmdRotationX, float? hmdRotationY, float? hmdRotationZ)
     {
